requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.1
bidict==0.19.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
comtypes==1.1.7
confusable-homoglyphs==3.2.0
cryptography==3.1.1
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.1
django-allauth==0.42.0
django-crispy-forms==1.9.0
django-filter==2.3.0
django-model-utils==4.0.0
django-quiz-app==0.5.1
django-registration==3.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.11.1
filelock==3.0.12
future==0.18.2
geocoder==1.38.1
idna==2.10
isort==4.3.21
jsonfield==3.1.0
lazy-object-proxy==1.4.3
loc==0.0.2
Markdown==3.2.2
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.3
pyparsing==2.4.7
pypiwin32==223
python-decouple==3.3
python3-openid==3.2.0
pyttsx3==2.90
pytz==2020.1
pywin32==228
ratelim==0.1.6
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.3.1
toml==0.10.1
tox==3.20.0
urllib3==1.25.9
virtualenv==20.0.31
wikipedia==1.4.0
wrapt==1.12.1
